+++ NEWS.md
@@ -1,5 +1,8 @@
 # CHANGES IN RRPP 2.0.2.999
 
+### NEW FEATURES
++ New function, `QRforX`, which produces consistent QR decomposition results despite differences between`base::qr` and `Matrix::qr` functions.
+
 ### BUG FIXES
 + Fixed an important issue with `removeRedundant` and `getRank` to properly remove axes based on a `base::qr` pivot strategy, but still avoiding the making of large, dense matrices before using `base::qr`.